+++ bindings/nostr-sdk-ffi/src/protocol/event/builder.rs
@@ -14,7 +14,7 @@ use super::{Event, EventId, Kind};
 use crate::error::Result;
 use crate::protocol::event::{PublicKey, Tag, Timestamp, UnsignedEvent};
 use crate::protocol::key::Keys;
-use crate::protocol::nips::nip01::Coordinate;
+use crate::protocol::nips::nip01::{Coordinate, Metadata};
 use crate::protocol::nips::nip15::{ProductData, StallData};
 use crate::protocol::nips::nip34::{GitIssue, GitPatch, GitRepositoryAnnouncement};
 use crate::protocol::nips::nip46::NostrConnectMessage;
@@ -28,7 +28,7 @@ use crate::protocol::nips::nip90::JobFeedbackData;
 use crate::protocol::nips::nip94::FileMetadata;
 use crate::protocol::nips::nip98::HttpData;
 use crate::protocol::signer::NostrSigner;
-use crate::protocol::types::{Contact, ImageDimensions, Metadata};
+use crate::protocol::types::{Contact, ImageDimensions};
 use crate::util::parse_optional_relay_url;
 
 #[derive(Debug, Clone, PartialEq, Eq, Object)]