+++ bindings/nostr-sdk-ffi/src/protocol/types/mod.rs
@@ -4,10 +4,8 @@
 
 pub mod contact;
 pub mod image;
-pub mod metadata;
 pub mod time;
 
 pub use self::contact::Contact;
 pub use self::image::ImageDimensions;
-pub use self::metadata::Metadata;
 pub use self::time::Timestamp;