+++ crates/nostr/src/lib.rs
@@ -65,11 +65,13 @@ pub use self::key::{Keys, PublicKey, SecretKey};
 #[doc(hidden)]
 pub use self::message::{ClientMessage, RawRelayMessage, RelayMessage, SubscriptionId};
 #[doc(hidden)]
+pub use self::nips::nip01::Metadata;
+#[doc(hidden)]
 pub use self::nips::nip19::{FromBech32, ToBech32};
 #[doc(hidden)]
 pub use self::signer::{NostrSigner, SignerError};
 #[doc(hidden)]
-pub use self::types::{Contact, ImageDimensions, Metadata, RelayUrl, Timestamp, TryIntoUrl, Url};
+pub use self::types::{Contact, ImageDimensions, RelayUrl, Timestamp, TryIntoUrl, Url};
 #[doc(hidden)]
 pub use self::util::JsonUtil;
 #[doc(hidden)]