+++ crates/nostr/src/nips/nip01.rs
@@ -7,17 +7,26 @@
 //! <https://github.com/nostr-protocol/nips/blob/master/01.md>
 
 use alloc::borrow::ToOwned;
+#[cfg(not(feature = "std"))]
+use alloc::collections::BTreeMap as AllocMap;
 use alloc::string::String;
 use alloc::vec::Vec;
 use core::fmt;
 use core::num::ParseIntError;
 use core::str::FromStr;
+#[cfg(feature = "std")]
+use std::collections::HashMap as AllocMap;
+
+use serde::de::{Deserializer, MapAccess, Visitor};
+use serde::ser::{SerializeMap, Serializer};
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
 
 use super::nip19::FromBech32;
 use super::nip21::NostrURI;
 use crate::event::id;
-use crate::types::RelayUrl;
-use crate::{key, Filter, Kind, PublicKey, Tag, TagStandard};
+use crate::types::{RelayUrl, Url};
+use crate::{key, Filter, JsonUtil, Kind, PublicKey, Tag, TagStandard};
 
 /// Raw Event error
 #[derive(Debug, PartialEq, Eq)]
@@ -208,9 +217,259 @@ impl FromStr for Coordinate {
     }
 }
 
+/// Metadata
+#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Metadata {
+    /// Name
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub name: Option<String>,
+    /// Display name
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub display_name: Option<String>,
+    /// Description
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub about: Option<String>,
+    /// Website url
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub website: Option<String>,
+    /// Picture url
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub picture: Option<String>,
+    /// Banner url
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub banner: Option<String>,
+    /// NIP05 (ex. name@example.com)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub nip05: Option<String>,
+    /// LNURL
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub lud06: Option<String>,
+    /// Lightning Address
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default)]
+    pub lud16: Option<String>,
+    /// Custom fields
+    #[serde(
+        flatten,
+        serialize_with = "serialize_custom_fields",
+        deserialize_with = "deserialize_custom_fields"
+    )]
+    #[serde(default)]
+    pub custom: AllocMap<String, Value>,
+}
+
+impl Metadata {
+    /// New empty [`Metadata`]
+    #[inline]
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Set name
+    pub fn name<S>(self, name: S) -> Self
+    where
+        S: Into<String>,
+    {
+        Self {
+            name: Some(name.into()),
+            ..self
+        }
+    }
+
+    /// Set display name
+    pub fn display_name<S>(self, display_name: S) -> Self
+    where
+        S: Into<String>,
+    {
+        Self {
+            display_name: Some(display_name.into()),
+            ..self
+        }
+    }
+
+    /// Set about
+    pub fn about<S>(self, about: S) -> Self
+    where
+        S: Into<String>,
+    {
+        Self {
+            about: Some(about.into()),
+            ..self
+        }
+    }
+
+    /// Set website
+    pub fn website(self, url: Url) -> Self {
+        Self {
+            website: Some(url.into()),
+            ..self
+        }
+    }
+
+    /// Set picture
+    pub fn picture(self, url: Url) -> Self {
+        Self {
+            picture: Some(url.into()),
+            ..self
+        }
+    }
+
+    /// Set banner
+    pub fn banner(self, url: Url) -> Self {
+        Self {
+            banner: Some(url.into()),
+            ..self
+        }
+    }
+
+    /// Set nip05
+    pub fn nip05<S>(self, nip05: S) -> Self
+    where
+        S: Into<String>,
+    {
+        Self {
+            nip05: Some(nip05.into()),
+            ..self
+        }
+    }
+
+    /// Set lud06 (LNURL)
+    pub fn lud06<S>(self, lud06: S) -> Self
+    where
+        S: Into<String>,
+    {
+        Self {
+            lud06: Some(lud06.into()),
+            ..self
+        }
+    }
+
+    /// Set lud16 (Lightning Address)
+    pub fn lud16<S>(self, lud16: S) -> Self
+    where
+        S: Into<String>,
+    {
+        Self {
+            lud16: Some(lud16.into()),
+            ..self
+        }
+    }
+
+    /// Set custom metadata field
+    pub fn custom_field<K, S>(mut self, field_name: K, value: S) -> Self
+    where
+        K: Into<String>,
+        S: Into<Value>,
+    {
+        self.custom.insert(field_name.into(), value.into());
+        self
+    }
+}
+
+impl JsonUtil for Metadata {
+    type Err = serde_json::Error;
+}
+
+fn serialize_custom_fields<S>(
+    custom_fields: &AllocMap<String, Value>,
+    serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    let mut map = serializer.serialize_map(Some(custom_fields.len()))?;
+    for (field_name, value) in custom_fields {
+        map.serialize_entry(field_name, value)?;
+    }
+    map.end()
+}
+
+fn deserialize_custom_fields<'de, D>(deserializer: D) -> Result<AllocMap<String, Value>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    struct GenericTagsVisitor;
+
+    impl<'de> Visitor<'de> for GenericTagsVisitor {
+        type Value = AllocMap<String, Value>;
+
+        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+            formatter.write_str("map where keys are strings and values are valid json")
+        }
+
+        fn visit_map<M>(self, mut map: M) -> Result<Self::Value, M::Error>
+        where
+            M: MapAccess<'de>,
+        {
+            #[cfg(not(feature = "std"))]
+            let mut custom_fields: AllocMap<String, Value> = AllocMap::new();
+            #[cfg(feature = "std")]
+            let mut custom_fields: AllocMap<String, Value> =
+                AllocMap::with_capacity(map.size_hint().unwrap_or_default());
+            while let Some(field_name) = map.next_key::<String>()? {
+                if let Ok(value) = map.next_value::<Value>() {
+                    custom_fields.insert(field_name, value);
+                }
+            }
+            Ok(custom_fields)
+        }
+    }
+
+    deserializer.deserialize_map(GenericTagsVisitor)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn test_deserialize_metadata() {
+        let content = r#"{"name":"myname","about":"Description","display_name":""}"#;
+        let metadata = Metadata::from_json(content).unwrap();
+        assert_eq!(
+            metadata,
+            Metadata::new()
+                .name("myname")
+                .about("Description")
+                .display_name("")
+        );
+
+        let content = r#"{"name":"myname","about":"Description","displayName":"Jack"}"#;
+        let metadata = Metadata::from_json(content).unwrap();
+        assert_eq!(
+            metadata,
+            Metadata::new()
+                .name("myname")
+                .about("Description")
+                .custom_field("displayName", "Jack")
+        );
+
+        let content = r#"{"lud16":"thesimplekid@cln.thesimplekid.com","nip05":"_@thesimplekid.com","display_name":"thesimplekid","about":"Wannabe open source dev","name":"thesimplekid","username":"thesimplekid","displayName":"thesimplekid","lud06":"","reactions":false,"damus_donation_v2":0}"#;
+        let metadata = Metadata::from_json(content).unwrap();
+        assert_eq!(
+            metadata,
+            Metadata::new()
+                .name("thesimplekid")
+                .display_name("thesimplekid")
+                .about("Wannabe open source dev")
+                .nip05("_@thesimplekid.com")
+                .lud06("")
+                .lud16("thesimplekid@cln.thesimplekid.com")
+                .custom_field("username", "thesimplekid")
+                .custom_field("displayName", "thesimplekid")
+                .custom_field("reactions", false)
+                .custom_field("damus_donation_v2", 0)
+        );
+        assert_eq!(metadata, Metadata::from_json(metadata.as_json()).unwrap());
+    }
 
     #[test]
     fn parse_valid_coordinate() {