+++ crates/nostr/src/types/mod.rs
@@ -9,12 +9,10 @@
 
 pub mod contact;
 pub mod image;
-pub mod metadata;
 pub mod time;
 pub mod url;
 
 pub use self::contact::*;
 pub use self::image::*;
-pub use self::metadata::*;
 pub use self::time::*;
 pub use self::url::*;